requirements.txt
Flask
Flask-Login
Flask-MySQLDB
python-dotenv
openai
google-generativeai
langchain
PyPDF2
scikit-learn
